move optimize/2 to the api module
